Merge in SLWFAT/simplelink_wi-fi_demos from bugfix/SLWIFI_SWAPPS-443-current-submodule-points-to-9.22.0.15-instead-of-latest-sdk to master

* commit 'd403703d45b09de7e0654235d3168bedfd8ecd92':
  Revert "freertosconfig change"
  freertosconfig change
  added GCC build conf for all examples. fixed any build issues
  no need for HTML
  removal of keys and SSID info
  build test script update to support makefiles only
  update gitignore for build artifacts
  update correct versioning in readme
  Fix Submodule issue

+++ src/freertos/cloud/aws/aws_iot_telemetry.c
@@ -339,7 +339,7 @@ uint32_t prvConnectToServerWithBackoffRetries( const char * pcHostName,
     TlsTransportStatus_t xNetworkStatus;
     BackoffAlgorithmStatus_t xBackoffAlgStatus = BackoffAlgorithmSuccess;
     BackoffAlgorithmContext_t xReconnectParams;
-    uint16_t usNextRetryBackOff = 0U;
+    uint32_t usNextRetryBackOff = 0U;
 
     /* Initialize reconnect attempts and interval. */
     BackoffAlgorithm_InitializeParams( &xReconnectParams,
